requirements.txt
boto3
langchain
openai
python-dotenv